move not-found in components/common
